requirements.txt
ipykernel
pyyaml
fastapi
uvicorn[standard]
pydantic
pydantic-settings
httpx
numpy
redis
asyncpg
sqlalchemy[asyncio]
alembic
tenacity
structlog
prometheus-client
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
python-jose[cryptography]
passlib[bcrypt]
python-multipart
tiktoken
sentence-transformers
faiss-cpu
aiofiles
pytest
pytest-asyncio
locust
llama_index
langgraph
langchain
langchain_text_splitters
qdrant_client
fitz
easyocr
python-docx
python-pptx
pytesseract
ragas
datasets
